+++ client/user_test.go
@@ -1,8 +1,1 @@
 package client
-
-import (
-	"testing"
-)
-
-func TestClientUserHasStartedReport(t *testing.T) {
-}
