+++ models/meeting_test.go
@@ -1,4 +1,4 @@
-package client
+package models
 
 import (
 	"testing"
