+++ models/question.go
@@ -1,0 +1,10 @@
+package models
+
+// Question a question to ask a user
+// Color is the attachment color
+// If options are given, the question will have a drop down
+type Question struct {
+	Text    string   `yaml:"text"`
+	Color   string   `yaml:"color"`
+	Options []string `yaml:"options,omitempty"`
+}
